// pharmacy-ui/src/app/services/models/page-response-product-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProductResponse } from './product-response';
export interface PageResponseProductResponse {
  content?: Array<ProductResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}